Refactor thresholding to segmentation
We decided to move the thresholding functions to another place, because
the filters module is a weird fit for them. `skimage.segmentation`
seemed like a good fit. In general, segmentation is a simple form  or
outcome of thresholding.

In the process improve on the API itself. E.g., grouping local
thresholding functions by name.
