+++ skimage/filters/thresholding.py
@@ -1,23 +1,10 @@
-import inspect
-import itertools
-import math
-from collections import OrderedDict
-from collections.abc import Iterable
-
 import numpy as np
 from scipy import ndimage as ndi
 
-from .._shared.filters import gaussian
-from .._shared.utils import _supported_float_type, warn
+from .._shared.utils import deprecate_func
 from .._shared.version_requirements import require
-from ..exposure import histogram
-from ..filters._multiotsu import (
-    _get_multiotsu_thresh_indices,
-    _get_multiotsu_thresh_indices_lut,
-)
-from ..transform import integral_image
-from ..util import dtype_limits
-from ._sparse import _correlate_sparse, _validate_window_size
+from ..segmentation import _thresholding_global, _thresholding_local
+
 
 __all__ = [
     'try_all_threshold',
@@ -36,75 +23,11 @@ __all__ = [
 ]
 
 
-def _try_all(image, methods=None, figsize=None, num_cols=2, verbose=True):
-    """Returns a figure comparing the outputs of different methods.
-
-    Parameters
-    ----------
-    image : (M, N) ndarray
-        Input image.
-    methods : dict, optional
-        Names and associated functions.
-        Functions must take and return an image.
-    figsize : tuple, optional
-        Figure size (in inches).
-    num_cols : int, optional
-        Number of columns.
-    verbose : bool, optional
-        Print function name for each method.
-
-    Returns
-    -------
-    fig, ax : tuple
-        Matplotlib figure and axes.
-    """
-    from matplotlib import pyplot as plt
-
-    # Compute the image histogram for better performances
-    nbins = 256  # Default in threshold functions
-    hist = histogram(image.reshape(-1), nbins, source_range='image')
-
-    # Handle default value
-    methods = methods or {}
-
-    num_rows = math.ceil((len(methods) + 1.0) / num_cols)
-    fig, ax = plt.subplots(
-        num_rows, num_cols, figsize=figsize, sharex=True, sharey=True
-    )
-    ax = ax.reshape(-1)
-
-    ax[0].imshow(image, cmap=plt.cm.gray)
-    ax[0].set_title('Original')
-
-    i = 1
-    for name, func in methods.items():
-        # Use precomputed histogram for supporting functions
-        sig = inspect.signature(func)
-        _kwargs = dict(hist=hist) if 'hist' in sig.parameters else {}
-
-        ax[i].set_title(name)
-        try:
-            ax[i].imshow(func(image, **_kwargs), cmap=plt.cm.gray)
-        except Exception as e:
-            ax[i].text(
-                0.5,
-                0.5,
-                f"{type(e).__name__}",
-                ha="center",
-                va="center",
-                transform=ax[i].transAxes,
-            )
-        i += 1
-        if verbose:
-            print(func.__orifunc__)
-
-    for a in ax:
-        a.axis('off')
-
-    fig.tight_layout()
-    return fig, ax
-
-
+@deprecate_func(
+    deprecated_version="0.26",
+    removed_version="2.0 (or later)",
+    hint="Use `skimage.segmentation.threshold_try_global` instead",
+)
 @require("matplotlib", ">=3.3")
 def try_all_threshold(image, figsize=(8, 5), verbose=True):
     """Returns a figure comparing the outputs of different thresholding methods.
@@ -143,37 +66,16 @@ def try_all_threshold(image, figsize=(8, 5), verbose=True):
     >>> from skimage.data import text
     >>> fig, ax = try_all_threshold(text(), figsize=(10, 6), verbose=False)
     """
-
-    def thresh(func):
-        """
-        A wrapper function to return a thresholded image.
-        """
-
-        def wrapper(im):
-            return im > func(im)
-
-        try:
-            wrapper.__orifunc__ = func.__orifunc__
-        except AttributeError:
-            wrapper.__orifunc__ = func.__module__ + '.' + func.__name__
-        return wrapper
-
-    # Global algorithms.
-    methods = OrderedDict(
-        {
-            'Isodata': thresh(threshold_isodata),
-            'Li': thresh(threshold_li),
-            'Mean': thresh(threshold_mean),
-            'Minimum': thresh(threshold_minimum),
-            'Otsu': thresh(threshold_otsu),
-            'Triangle': thresh(threshold_triangle),
-            'Yen': thresh(threshold_yen),
-        }
+    return _thresholding_global.threshold_try_global(
+        image=image, figsize=figsize, verbose=verbose
     )
 
-    return _try_all(image, figsize=figsize, methods=methods, verbose=verbose)
 
-
+@deprecate_func(
+    deprecated_version="0.26",
+    removed_version="2.0 (or later)",
+    hint="Use `skimage.segmentation.threshold_local` instead",
+)
 def threshold_local(
     image, block_size=3, method='gaussian', offset=0, mode='reflect', param=None, cval=0
 ):
@@ -239,100 +141,22 @@ def threshold_local(
     ...                                         param=func)
 
     """
-
-    if np.isscalar(block_size):
-        block_size = (block_size,) * image.ndim
-    elif len(block_size) != image.ndim:
-        raise ValueError("len(block_size) must equal image.ndim.")
-    block_size = tuple(block_size)
-    if any(b % 2 == 0 for b in block_size):
-        raise ValueError(
-            f'block_size must be odd! Given block_size '
-            f'{block_size} contains even values.'
-        )
-    float_dtype = _supported_float_type(image.dtype)
-    image = image.astype(float_dtype, copy=False)
-    thresh_image = np.zeros(image.shape, dtype=float_dtype)
-    if method == 'generic':
-        ndi.generic_filter(
-            image, param, block_size, output=thresh_image, mode=mode, cval=cval
-        )
-    elif method == 'gaussian':
-        if param is None:
-            # automatically determine sigma which covers > 99% of distribution
-            sigma = tuple([(b - 1) / 6.0 for b in block_size])
-        else:
-            sigma = param
-        gaussian(image, sigma=sigma, out=thresh_image, mode=mode, cval=cval)
-    elif method == 'mean':
-        ndi.uniform_filter(image, block_size, output=thresh_image, mode=mode, cval=cval)
-    elif method == 'median':
-        ndi.median_filter(image, block_size, output=thresh_image, mode=mode, cval=cval)
-    else:
-        raise ValueError(
-            "Invalid method specified. Please use `generic`, "
-            "`gaussian`, `mean`, or `median`."
-        )
-
-    return thresh_image - offset
+    return _thresholding_local.threshold_local(
+        image=image,
+        block_size=block_size,
+        method=method,
+        offset=offset,
+        mode=mode,
+        param=param,
+        cval=cval,
+    )
 
 
-def _validate_image_histogram(image, hist, nbins=None, normalize=False):
-    """Ensure that either image or hist were given, return valid histogram.
-
-    If hist is given, image is ignored.
-
-    Parameters
-    ----------
-    image : array or None
-        Grayscale image.
-    hist : array, 2-tuple of array, or None
-        Histogram, either a 1D counts array, or an array of counts together
-        with an array of bin centers.
-    nbins : int, optional
-        The number of bins with which to compute the histogram, if `hist` is
-        None.
-    normalize : bool
-        If hist is not given, it will be computed by this function. This
-        parameter determines whether the computed histogram is normalized
-        (i.e. entries sum up to 1) or not.
-
-    Returns
-    -------
-    counts : 1D array of float
-        Each element is the number of pixels falling in each intensity bin.
-    bin_centers : 1D array
-        Each element is the value corresponding to the center of each intensity
-        bin.
-
-    Raises
-    ------
-    ValueError : if image and hist are both None
-    """
-    if image is None and hist is None:
-        raise Exception("Either image or hist must be provided.")
-
-    if hist is not None:
-        if isinstance(hist, (tuple, list)):
-            counts, bin_centers = hist
-        else:
-            counts = hist
-            bin_centers = np.arange(counts.size)
-
-        if counts[0] == 0 or counts[-1] == 0:
-            # Trim histogram from both ends by removing starting and
-            # ending zeroes as in histogram(..., source_range="image")
-            cond = counts > 0
-            start = np.argmax(cond)
-            end = cond.size - np.argmax(cond[::-1])
-            counts, bin_centers = counts[start:end], bin_centers[start:end]
-    else:
-        counts, bin_centers = histogram(
-            image.reshape(-1), nbins, source_range='image', normalize=normalize
-        )
-    return counts.astype('float32', copy=False), bin_centers
-
-
+@deprecate_func(
+    deprecated_version="0.26",
+    removed_version="2.0 (or later)",
+    hint="Use `skimage.segmentation.threshold_otsu` instead",
+)
 def threshold_otsu(image=None, nbins=256, *, hist=None):
     """Return threshold value based on Otsu's method.
 
@@ -373,40 +197,14 @@ def threshold_otsu(image=None, nbins=256, *, hist=None):
     -----
     The input image must be grayscale.
     """
-    if image is not None and image.ndim > 2 and image.shape[-1] in (3, 4):
-        warn(
-            f'threshold_otsu is expected to work correctly only for '
-            f'grayscale images; image shape {image.shape} looks like '
-            f'that of an RGB image.'
-        )
-
-    # Check if the image has more than one intensity value; if not, return that
-    # value
-    if image is not None:
-        first_pixel = image.reshape(-1)[0]
-        if np.all(image == first_pixel):
-            return first_pixel
-
-    counts, bin_centers = _validate_image_histogram(image, hist, nbins)
-
-    # class probabilities for all possible thresholds
-    weight1 = np.cumsum(counts)
-    weight2 = np.cumsum(counts[::-1])[::-1]
-    # class means for all possible thresholds
-    mean1 = np.cumsum(counts * bin_centers) / weight1
-    mean2 = (np.cumsum((counts * bin_centers)[::-1]) / weight2[::-1])[::-1]
-
-    # Clip ends to align class 1 and class 2 variables:
-    # The last value of ``weight1``/``mean1`` should pair with zero values in
-    # ``weight2``/``mean2``, which do not exist.
-    variance12 = weight1[:-1] * weight2[1:] * (mean1[:-1] - mean2[1:]) ** 2
-
-    idx = np.argmax(variance12)
-    threshold = bin_centers[idx]
-
-    return threshold
+    return _thresholding_global.threshold_otsu(image=image, nbins=nbins, hist=hist)
 
 
+@deprecate_func(
+    deprecated_version="0.26",
+    removed_version="2.0 (or later)",
+    hint="Use `skimage.segmentation.threshold_yen` instead",
+)
 def threshold_yen(image=None, nbins=256, *, hist=None):
     """Return threshold value based on Yen's method.
     Either image or hist must be provided. In case hist is given, the actual
@@ -448,25 +246,14 @@ def threshold_yen(image=None, nbins=256, *, hist=None):
     >>> thresh = threshold_yen(image)
     >>> binary = image <= thresh
     """
-    counts, bin_centers = _validate_image_histogram(image, hist, nbins)
-
-    # On blank images (e.g. filled with 0) with int dtype, `histogram()`
-    # returns ``bin_centers`` containing only one value. Speed up with it.
-    if bin_centers.size == 1:
-        return bin_centers[0]
-
-    # Calculate probability mass function
-    pmf = counts.astype('float32', copy=False) / counts.sum()
-    P1 = np.cumsum(pmf)  # Cumulative normalized histogram
-    P1_sq = np.cumsum(pmf**2)
-    # Get cumsum calculated from end of squared array:
-    P2_sq = np.cumsum(pmf[::-1] ** 2)[::-1]
-    # P2_sq indexes is shifted +1. I assume, with P1[:-1] it's help avoid
-    # '-inf' in crit. ImageJ Yen implementation replaces those values by zero.
-    crit = np.log(((P1_sq[:-1] * P2_sq[1:]) ** -1) * (P1[:-1] * (1.0 - P1[:-1])) ** 2)
-    return bin_centers[crit.argmax()]
+    return _thresholding_global.threshold_yen(image=image, nbins=nbins, hist=hist)
 
 
+@deprecate_func(
+    deprecated_version="0.26",
+    removed_version="2.0 (or later)",
+    hint="Use `skimage.segmentation.threshold_isodata` instead",
+)
 def threshold_isodata(image=None, nbins=256, return_all=False, *, hist=None):
     """Return threshold value(s) based on ISODATA method.
 
@@ -527,61 +314,9 @@ def threshold_isodata(image=None, nbins=256, return_all=False, *, hist=None):
     >>> thresh = threshold_isodata(image)
     >>> binary = image > thresh
     """
-    counts, bin_centers = _validate_image_histogram(image, hist, nbins)
-
-    # image only contains one unique value
-    if len(bin_centers) == 1:
-        if return_all:
-            return bin_centers
-        else:
-            return bin_centers[0]
-
-    counts = counts.astype('float32', copy=False)
-
-    # csuml and csumh contain the count of pixels in that bin or lower, and
-    # in all bins strictly higher than that bin, respectively
-    csuml = np.cumsum(counts)
-    csumh = csuml[-1] - csuml
-
-    # intensity_sum contains the total pixel intensity from each bin
-    intensity_sum = counts * bin_centers
-
-    # l and h contain average value of all pixels in that bin or lower, and
-    # in all bins strictly higher than that bin, respectively.
-    # Note that since exp.histogram does not include empty bins at the low or
-    # high end of the range, csuml and csumh are strictly > 0, except in the
-    # last bin of csumh, which is zero by construction.
-    # So no worries about division by zero in the following lines, except
-    # for the last bin, but we can ignore that because no valid threshold
-    # can be in the top bin.
-    # To avoid the division by zero, we simply skip over the last element in
-    # all future computation.
-    csum_intensity = np.cumsum(intensity_sum)
-    lower = csum_intensity[:-1] / csuml[:-1]
-    higher = (csum_intensity[-1] - csum_intensity[:-1]) / csumh[:-1]
-
-    # isodata finds threshold values that meet the criterion t = (l + m)/2
-    # where l is the mean of all pixels <= t and h is the mean of all pixels
-    # > t, as calculated above. So we are looking for places where
-    # (l + m) / 2 equals the intensity value for which those l and m figures
-    # were calculated -- which is, of course, the histogram bin centers.
-    # We only require this equality to be within the precision of the bin
-    # width, of course.
-    all_mean = (lower + higher) / 2.0
-    bin_width = bin_centers[1] - bin_centers[0]
-
-    # Look only at thresholds that are below the actual all_mean value,
-    # for consistency with the threshold being included in the lower pixel
-    # group. Otherwise, can get thresholds that are not actually fixed-points
-    # of the isodata algorithm. For float images, this matters less, since
-    # there really can't be any guarantees anymore anyway.
-    distances = all_mean - bin_centers[:-1]
-    thresholds = bin_centers[:-1][(distances >= 0) & (distances < bin_width)]
-
-    if return_all:
-        return thresholds
-    else:
-        return thresholds[0]
+    return _thresholding_global.threshold_isodata(
+        image=image, nbins=nbins, return_all=return_all, hist=hist
+    )
 
 
 # Computing a histogram using np.histogram on a uint8 image with bins=256
@@ -639,6 +374,11 @@ def _cross_entropy(image, threshold, bins=_DEFAULT_ENTROPY_BINS):
     return nu
 
 
+@deprecate_func(
+    deprecated_version="0.26",
+    removed_version="2.0 (or later)",
+    hint="Use `skimage.segmentation.threshold_li` instead",
+)
 def threshold_li(image, *, tolerance=None, initial_guess=None, iter_callback=None):
     """Compute threshold value by Li's iterative Minimum Cross Entropy method.
 
@@ -691,107 +431,19 @@ def threshold_li(image, *, tolerance=None, initial_guess=None, iter_callback=Non
     >>> thresh = threshold_li(image)
     >>> binary = image > thresh
     """
-    # Remove nan:
-    image = image[~np.isnan(image)]
-    if image.size == 0:
-        return np.nan
-
-    # Make sure image has more than one value; otherwise, return that value
-    # This works even for np.inf
-    if np.all(image == image.flat[0]):
-        return image.flat[0]
-
-    # At this point, the image only contains np.inf, -np.inf, or valid numbers
-    image = image[np.isfinite(image)]
-    # if there are no finite values in the image, return 0. This is because
-    # at this point we *know* that there are *both* inf and -inf values,
-    # because inf == inf evaluates to True. We might as well separate them.
-    if image.size == 0:
-        return 0.0
-
-    # Li's algorithm requires positive image (because of log(mean))
-    image_min = np.min(image)
-    image -= image_min
-    if image.dtype.kind in 'iu':
-        tolerance = tolerance or 0.5
-    else:
-        tolerance = tolerance or np.min(np.diff(np.unique(image))) / 2
-
-    # Initial estimate for iteration. See "initial_guess" in the parameter list
-    if initial_guess is None:
-        t_next = np.mean(image)
-    elif callable(initial_guess):
-        t_next = initial_guess(image)
-    elif np.isscalar(initial_guess):  # convert to new, positive image range
-        t_next = initial_guess - float(image_min)
-        image_max = np.max(image) + image_min
-        if not 0 < t_next < np.max(image):
-            msg = (
-                f'The initial guess for threshold_li must be within the '
-                f'range of the image. Got {initial_guess} for image min '
-                f'{image_min} and max {image_max}.'
-            )
-            raise ValueError(msg)
-        t_next = image.dtype.type(t_next)
-    else:
-        raise TypeError(
-            'Incorrect type for `initial_guess`; should be '
-            'a floating point value, or a function mapping an '
-            'array to a floating point value.'
-        )
-
-    # initial value for t_curr must be different from t_next by at
-    # least the tolerance. Since the image is positive, we ensure this
-    # by setting to a large-enough negative number
-    t_curr = -2 * tolerance
-
-    # Callback on initial iterations
-    if iter_callback is not None:
-        iter_callback(t_next + image_min)
-
-    # Stop the iterations when the difference between the
-    # new and old threshold values is less than the tolerance
-    # or if the background mode has only one value left,
-    # since log(0) is not defined.
-
-    if image.dtype.kind in 'iu':
-        hist, bin_centers = histogram(image.reshape(-1), source_range='image')
-        hist = hist.astype('float32', copy=False)
-        while abs(t_next - t_curr) > tolerance:
-            t_curr = t_next
-            foreground = bin_centers > t_curr
-            background = ~foreground
-
-            mean_fore = np.average(bin_centers[foreground], weights=hist[foreground])
-            mean_back = np.average(bin_centers[background], weights=hist[background])
-
-            if mean_back == 0:
-                break
-
-            t_next = (mean_back - mean_fore) / (np.log(mean_back) - np.log(mean_fore))
-
-            if iter_callback is not None:
-                iter_callback(t_next + image_min)
-
-    else:
-        while abs(t_next - t_curr) > tolerance:
-            t_curr = t_next
-            foreground = image > t_curr
-            mean_fore = np.mean(image[foreground])
-            mean_back = np.mean(image[~foreground])
-
-            if mean_back == 0.0:
-                break
-
-            t_next = (mean_back - mean_fore) / (np.log(mean_back) - np.log(mean_fore))
-
-            if iter_callback is not None:
-                iter_callback(t_next + image_min)
-
-    threshold = t_next + image_min
-    return threshold
+    return _thresholding_global.threshold_li(
+        image=image,
+        tolerance=tolerance,
+        initial_guess=initial_guess,
+        iter_callback=iter_callback,
+    )
 
 
+@deprecate_func(
+    deprecated_version="0.26",
+    removed_version="2.0 (or later)",
+    hint="Use `skimage.segmentation.threshold_minimum` instead",
+)
 def threshold_minimum(image=None, nbins=256, max_num_iter=10000, *, hist=None):
     """Return threshold value based on minimum method.
 
@@ -844,45 +496,16 @@ def threshold_minimum(image=None, nbins=256, max_num_iter=10000, *, hist=None):
     >>> thresh = threshold_minimum(image)
     >>> binary = image > thresh
     """
-
-    def find_local_maxima_idx(hist):
-        # We can't use scipy.signal.argrelmax
-        # as it fails on plateaus
-        maximum_idxs = list()
-        direction = 1
-
-        for i in range(hist.shape[0] - 1):
-            if direction > 0:
-                if hist[i + 1] < hist[i]:
-                    direction = -1
-                    maximum_idxs.append(i)
-            else:
-                if hist[i + 1] > hist[i]:
-                    direction = 1
-
-        return maximum_idxs
-
-    counts, bin_centers = _validate_image_histogram(image, hist, nbins)
-
-    smooth_hist = counts.astype('float32', copy=False)
-
-    for counter in range(max_num_iter):
-        smooth_hist = ndi.uniform_filter1d(smooth_hist, 3)
-        maximum_idxs = find_local_maxima_idx(smooth_hist)
-        if len(maximum_idxs) < 3:
-            break
-
-    if len(maximum_idxs) != 2:
-        raise RuntimeError('Unable to find two maxima in histogram')
-    elif counter == max_num_iter - 1:
-        raise RuntimeError('Maximum iteration reached for histogram' 'smoothing')
-
-    # Find the lowest point between the maxima
-    threshold_idx = np.argmin(smooth_hist[maximum_idxs[0] : maximum_idxs[1] + 1])
-
-    return bin_centers[maximum_idxs[0] + threshold_idx]
+    return _thresholding_global.threshold_minimum(
+        image=image, nbins=nbins, max_num_iter=max_num_iter, hist=hist
+    )
 
 
+@deprecate_func(
+    deprecated_version="0.26",
+    removed_version="2.0 (or later)",
+    hint="Use `skimage.segmentation.threshold_mean` instead",
+)
 def threshold_mean(image):
     """Return threshold value based on the mean of grayscale values.
 
@@ -911,9 +534,14 @@ def threshold_mean(image):
     >>> thresh = threshold_mean(image)
     >>> binary = image > thresh
     """
-    return np.mean(image)
+    return _thresholding_global.threshold_mean(image)
 
 
+@deprecate_func(
+    deprecated_version="0.26",
+    removed_version="2.0 (or later)",
+    hint="Use `skimage.segmentation.threshold_triangle` instead",
+)
 def threshold_triangle(image, nbins=256):
     """Return threshold value based on the triangle algorithm.
 
@@ -947,119 +575,14 @@ def threshold_triangle(image, nbins=256):
     >>> thresh = threshold_triangle(image)
     >>> binary = image > thresh
     """
-    # nbins is ignored for integer arrays
-    # so, we recalculate the effective nbins.
-    hist, bin_centers = histogram(image.reshape(-1), nbins, source_range='image')
-    nbins = len(hist)
-
-    # Find peak, lowest and highest gray levels.
-    arg_peak_height = np.argmax(hist)
-    peak_height = hist[arg_peak_height]
-    arg_low_level, arg_high_level = np.flatnonzero(hist)[[0, -1]]
-
-    if arg_low_level == arg_high_level:
-        # Image has constant intensity.
-        return image.ravel()[0]
-
-    # Flip is True if left tail is shorter.
-    flip = arg_peak_height - arg_low_level < arg_high_level - arg_peak_height
-    if flip:
-        hist = hist[::-1]
-        arg_low_level = nbins - arg_high_level - 1
-        arg_peak_height = nbins - arg_peak_height - 1
-
-    # If flip == True, arg_high_level becomes incorrect
-    # but we don't need it anymore.
-    del arg_high_level
-
-    # Set up the coordinate system.
-    width = arg_peak_height - arg_low_level
-    x1 = np.arange(width)
-    y1 = hist[x1 + arg_low_level]
-
-    # Normalize.
-    norm = np.sqrt(peak_height**2 + width**2)
-    peak_height /= norm
-    width /= norm
-
-    # Maximize the length.
-    # The ImageJ implementation includes an additional constant when calculating
-    # the length, but here we omit it as it does not affect the location of the
-    # minimum.
-    length = peak_height * x1 - width * y1
-    arg_level = np.argmax(length) + arg_low_level
-
-    if flip:
-        arg_level = nbins - arg_level - 1
-
-    return bin_centers[arg_level]
+    return _thresholding_global.threshold_triangle(image=image, nbins=nbins)
 
 
-def _mean_std(image, w):
-    """Return local mean and standard deviation of each pixel using a
-    neighborhood defined by a rectangular window size ``w``.
-    The algorithm uses integral images to speedup computation. This is
-    used by :func:`threshold_niblack` and :func:`threshold_sauvola`.
-
-    Parameters
-    ----------
-    image : (M, N[, ...]) ndarray
-        Grayscale input image.
-    w : int, or iterable of int
-        Window size specified as a single odd integer (3, 5, 7, …),
-        or an iterable of length ``image.ndim`` containing only odd
-        integers (e.g. ``(1, 5, 5)``).
-
-    Returns
-    -------
-    m : ndarray of float, same shape as ``image``
-        Local mean of the image.
-    s : ndarray of float, same shape as ``image``
-        Local standard deviation of the image.
-
-    References
-    ----------
-    .. [1] F. Shafait, D. Keysers, and T. M. Breuel, "Efficient
-           implementation of local adaptive thresholding techniques
-           using integral images." in Document Recognition and
-           Retrieval XV, (San Jose, USA), Jan. 2008.
-           :DOI:`10.1117/12.767755`
-    """
-
-    if not isinstance(w, Iterable):
-        w = (w,) * image.ndim
-    _validate_window_size(w)
-
-    float_dtype = _supported_float_type(image.dtype)
-    pad_width = tuple((k // 2 + 1, k // 2) for k in w)
-    padded = np.pad(image.astype(float_dtype, copy=False), pad_width, mode='reflect')
-
-    # Note: keep float64 integral images for accuracy. Outputs of
-    # _correlate_sparse can later be safely cast to float_dtype
-    integral = integral_image(padded, dtype=np.float64)
-    padded *= padded
-    integral_sq = integral_image(padded, dtype=np.float64)
-
-    # Create lists of non-zero kernel indices and values
-    kernel_indices = list(itertools.product(*tuple([(0, _w) for _w in w])))
-    kernel_values = [
-        (-1) ** (image.ndim % 2 != np.sum(indices) % 2) for indices in kernel_indices
-    ]
-
-    total_window_size = math.prod(w)
-    kernel_shape = tuple(_w + 1 for _w in w)
-    m = _correlate_sparse(integral, kernel_shape, kernel_indices, kernel_values)
-    m = m.astype(float_dtype, copy=False)
-    m /= total_window_size
-    g2 = _correlate_sparse(integral_sq, kernel_shape, kernel_indices, kernel_values)
-    g2 = g2.astype(float_dtype, copy=False)
-    g2 /= total_window_size
-    # Note: we use np.clip because g2 is not guaranteed to be greater than
-    # m*m when floating point error is considered
-    s = np.sqrt(np.clip(g2 - m * m, 0, None))
-    return m, s
-
-
+@deprecate_func(
+    deprecated_version="0.26",
+    removed_version="2.0 (or later)",
+    hint="Use `skimage.segmentation.threshold_local_niblack` instead",
+)
 def threshold_niblack(image, window_size=15, k=0.2):
     """Applies Niblack local threshold to an array.
 
@@ -1119,12 +642,18 @@ def threshold_niblack(image, window_size=15, k=0.2):
     >>> image = data.page()
     >>> threshold_image = threshold_niblack(image, window_size=7, k=0.1)
     """
-    m, s = _mean_std(image, window_size)
-    return m - k * s
+    return _thresholding_local.threshold_local_niblack(
+        image=image, window_size=window_size, k=k
+    )
 
 
+@deprecate_func(
+    deprecated_version="0.26",
+    removed_version="2.0 (or later)",
+    hint="Use `skimage.segmentation.threshold_local_sauvola` instead",
+)
 def threshold_sauvola(image, window_size=15, k=0.2, r=None):
-    """Applies Sauvola local threshold to an array. Sauvola is a
+    """Applies Sauvola lsocal threshold to an array. Sauvola is a
     modification of Niblack technique.
 
     In the original method a threshold T is calculated for every pixel
@@ -1176,11 +705,9 @@ def threshold_sauvola(image, window_size=15, k=0.2, r=None):
     >>> t_sauvola = threshold_sauvola(image, window_size=15, k=0.2)
     >>> binary_image = image > t_sauvola
     """
-    if r is None:
-        imin, imax = dtype_limits(image, clip_negative=False)
-        r = 0.5 * (imax - imin)
-    m, s = _mean_std(image, window_size)
-    return m * (1 + k * ((s / r) - 1))
+    return _thresholding_local.threshold_local_sauvola(
+        image=image, window_size=window_size, k=k, r=r
+    )
 
 
 def apply_hysteresis_threshold(image, low, high):
@@ -1230,6 +757,11 @@ def apply_hysteresis_threshold(image, low, high):
     return thresholded
 
 
+@deprecate_func(
+    deprecated_version="0.26",
+    removed_version="2.0 (or later)",
+    hint="Use `skimage.segmentation.threshold_multiotsu` instead",
+)
 def threshold_multiotsu(image=None, classes=3, nbins=256, *, hist=None):
     r"""Generate `classes`-1 threshold values to divide gray levels in `image`,
     following Otsu's method for multiple classes.
@@ -1301,39 +833,6 @@ def threshold_multiotsu(image=None, classes=3, nbins=256, *, hist=None):
     >>> regions = np.digitize(image, bins=thresholds)
     >>> regions_colorized = label2rgb(regions)
     """
-    if image is not None and image.ndim > 2 and image.shape[-1] in (3, 4):
-        warn(
-            f'threshold_multiotsu is expected to work correctly only for '
-            f'grayscale images; image shape {image.shape} looks like '
-            f'that of an RGB image.'
-        )
-
-    # calculating the histogram and the probability of each gray level.
-    prob, bin_centers = _validate_image_histogram(image, hist, nbins, normalize=True)
-    prob = prob.astype('float32', copy=False)
-
-    nvalues = np.count_nonzero(prob)
-    if nvalues < classes:
-        msg = (
-            f'After discretization into bins, the input image has '
-            f'only {nvalues} different values. It cannot be thresholded '
-            f'in {classes} classes. If there are more unique values '
-            f'before discretization, try increasing the number of bins '
-            f'(`nbins`).'
-        )
-        raise ValueError(msg)
-    elif nvalues == classes:
-        thresh_idx = np.flatnonzero(prob)[:-1]
-    else:
-        # Get threshold indices
-        try:
-            thresh_idx = _get_multiotsu_thresh_indices_lut(prob, classes - 1)
-        except MemoryError:
-            # Don't use LUT if the number of bins is too large (if the
-            # image is uint16 for example): in this case, the
-            # allocated memory is too large.
-            thresh_idx = _get_multiotsu_thresh_indices(prob, classes - 1)
-
-    thresh = bin_centers[thresh_idx]
-
-    return thresh
+    return _thresholding_global.threshold_multiotsu(
+        image=image, classes=classes, nbins=nbins, hist=hist
+    )
